requirements.txt
streamlit>=1.30.0
azure-identity>=1.15.0
openai>=1.70.0
